+++ src/main/kotlin/org/codefreak/codefreak/service/evaluation/EvaluationService.kt
@@ -24,6 +24,7 @@ import org.codefreak.codefreak.service.BaseService
 import org.codefreak.codefreak.service.EntityNotFoundException
 import org.codefreak.codefreak.service.IdeService
 import org.codefreak.codefreak.service.SubmissionService
+import org.codefreak.codefreak.service.TaskService
 import org.codefreak.codefreak.service.evaluation.runner.CommentRunner
 import org.codefreak.codefreak.service.file.FileService
 import org.codefreak.codefreak.util.PositionUtil
@@ -53,6 +54,9 @@ class EvaluationService : BaseService() {
 
   @Autowired
   private lateinit var answerService: AnswerService
+
+  @Autowired
+  private lateinit var taskService: TaskService
 
   @Autowired
   private lateinit var evaluationQueue: EvaluationQueue
@@ -232,5 +236,23 @@ class EvaluationService : BaseService() {
     val schema = JsonSchemaFactory.getInstance(SpecVersion.VersionFlag.V6).getSchema(runner.getOptionsSchema())
     val errors = schema.validate(objectMapper.valueToTree(definition.options))
     require(errors.isEmpty()) { "Runner options for ${definition.runnerName} are invalid: \n" + errors.joinToString("\n") { it.message } }
+  }
+
+  @Transactional
+  fun updateEvaluationStepDefinition(evaluationStepDefinition: EvaluationStepDefinition, title: String?, active: Boolean?, options: Map<String, Any>?): EvaluationStepDefinition {
+    title?.let {
+      evaluationStepDefinition.title = it
+    }
+    active?.let {
+      evaluationStepDefinition.active = it
+    }
+    options?.let {
+      evaluationStepDefinition.options = it
+    }
+    validateRunnerOptions(evaluationStepDefinition)
+    saveEvaluationStepDefinition(evaluationStepDefinition)
+    taskService.invalidateLatestEvaluations(evaluationStepDefinition.task)
+
+    return evaluationStepDefinition
   }
 }