+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * IDE can be disabled and custom images can be used (#606)
 * Tasks show the dates they were created and last updated (#617)
 * The task pool list and the 'add tasks to assignment' list are now sortable and filterable (#616)
+* The task pool can be exported and imported (#640)
 
 ### Changed